% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{CEM}
\alias{CEM}
\title{Clustering Matrix : One hot encoding}
\usage{
CEM(Z)
}
\arguments{
\item{Z}{a matrix N x K, with probabilities to belong of a cluster in rows for each observation.}
}
\value{
Z a matrix N x K One-Hot-Encoded by rows, where K is the number of clusters.
}
\description{
Clustering Matrix : One hot encoding
}
\examples{
Z <- matrix(rnorm(12),3,4)
Z_cem <- CEM(Z)
print(Z_cem)
}
