% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{diag_nulle}
\alias{diag_nulle}
\title{Diagonal coefficient to 0 on each slice given the 3rd dimension.}
\usage{
diag_nulle(A)
}
\arguments{
\item{A}{a array of dimension dim=c(N,N,V)}
}
\value{
A with 0 on each diagonal given the 3rd dimension.
}
\description{
Diagonal coefficient to 0 on each slice given the 3rd dimension.
}
