% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Initialization.R
\name{fit_SBM_per_layer}
\alias{fit_SBM_per_layer}
\title{SBM on each layer}
\usage{
fit_SBM_per_layer(A, silent = FALSE, ncores = 2)
}
\arguments{
\item{A}{an array of dim=c(N,N,V)}

\item{silent}{Boolean for verbose}

\item{ncores}{the number of cores used to parallelize the calculations of the various SBMs}
}
\value{
a list containing the parameters of each SBM applied to each view
}
\description{
SBM on each layer
}
