% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Simulation.R
\name{lab_switching}
\alias{lab_switching}
\title{Label Switching}
\usage{
lab_switching(Z, p_out = 0.1)
}
\arguments{
\item{Z}{a clustering vector}

\item{p_out}{a probability of perturbation for the clustering}
}
\value{
a perturbed clustering vector
}
\description{
This function can be used to perturb a clustering vector in order to randomly associate certain individuals with another cluster.
}
\examples{
Z <- sample(1:4,100,replace=TRUE)
p = 0.1
Z_pert <- lab_switching(Z,p)
table("Initial clustering" = Z,"Perturbed clustering" = Z_pert)
}
