% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{log_Softmax}
\alias{log_Softmax}
\title{log softmax of matrices (by row)}
\usage{
log_Softmax(log_X)
}
\arguments{
\item{log_X}{a matrix of log(X)}
}
\value{
X with log_softmax function applied on each row
}
\description{
log softmax of matrices (by row)
}
\examples{
set.seed(42)
X <- matrix(rnorm(15,mean=5),5,3)
log_X <- log(X)
X_softmax <-  log_Softmax(X)
}
