% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BayesianMixtureSBM.R
\name{Loss_BayesianMSBM}
\alias{Loss_BayesianMSBM}
\title{mimiSBM Evidence Lower BOund}
\usage{
Loss_BayesianMSBM(params)
}
\arguments{
\item{params}{a list of parameters of the model}
}
\value{
computation of the mimiSBM ELBO
}
\description{
mimiSBM Evidence Lower BOund
}
