% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{multinomial_lbeta_function}
\alias{multinomial_lbeta_function}
\title{Calculation of Log multinomial Beta value.}
\usage{
multinomial_lbeta_function(x)
}
\arguments{
\item{x}{a vector}
}
\value{
sum(lgamma(x[j])) - lgamma(sum(x))
}
\description{
Calculation of Log multinomial Beta value.
}
