% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{one_hot_errormachine}
\alias{one_hot_errormachine}
\title{One Hot Encoding with Error machine}
\usage{
one_hot_errormachine(Z, size = NULL)
}
\arguments{
\item{Z}{a vector of size N, where Z[i] value indicate the cluster membership of observation i.}

\item{size}{optional parameter, indicating the number of classes (avoid some empty class problems).}
}
\value{
Z a matrix N x K One-Hot-Encoded by rows, where K is the number of clusters.
}
\description{
One Hot Encoding with Error machine
}
\examples{
Z <- sample(1:4,10,replace=TRUE)
Z_OHE <- one_hot_errormachine(Z)
print(Z_OHE)
}
