% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Simulation.R
\name{partition_K_barre}
\alias{partition_K_barre}
\title{Create a link between final clustering and clustering per view component.}
\usage{
partition_K_barre(K_barre, K)
}
\arguments{
\item{K_barre}{Number of clusters in the final clustering}

\item{K}{Vector of size Q, indicate the number of clusters in each component. K[q] <= K_barre for all q}
}
\value{
cluster : a list of link between final clustering and clustering per view component.
}
\description{
Create a link between final clustering and clustering per view component.
}
