% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Visualisation.R
\name{plot_adjacency}
\alias{plot_adjacency}
\title{Plot adjacency matrices}
\usage{
plot_adjacency(A)
}
\arguments{
\item{A}{an array with dim=c(N,N,V).}
}
\value{
None
}
\description{
A function to plot each adjacency matrices defined by the thrid dimension of an array, and plot the sum of all theses matrices.
}
