% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Simulation.R
\name{rMSBM}
\alias{rMSBM}
\title{Simulate data from the mimiSBM generative model.}
\usage{
rMSBM(N, V, alpha_klq, pi_k, rho, sorted = TRUE, p_switch = NULL)
}
\arguments{
\item{N}{Number of individuals.}

\item{V}{Number of views.}

\item{alpha_klq}{array of component-connection probability (K,K,Q).}

\item{pi_k}{Vector of proportions of individuals across clusters.}

\item{rho}{Vector of proportion of views across components.}

\item{sorted}{Boolean for simulation reordering (clusters and components membership).}

\item{p_switch}{probability of label-switching, if NULL no perturbation between true clustering and the connectivity of individuals.}
}
\value{
list with the parameters of the simulation ($params), and
the simulations ($simulation).
}
\description{
Simulate data from the mimiSBM generative model.
}
