% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{sort_Z}
\alias{sort_Z}
\title{Sort the clustering matrix}
\usage{
sort_Z(Z)
}
\arguments{
\item{Z}{a matrix N x K, with probabilities to belong of a cluster in rows for each observation.}
}
\value{
a sorted matrix
}
\description{
Sort the clustering matrix
}
