% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{transpo}
\alias{transpo}
\title{Transposition of an array}
\usage{
transpo(A)
}
\arguments{
\item{A}{a array of dim= c(.,.,V)}
}
\value{
A_transposed, the transposed array according the third dimension
}
\description{
Transposition of an array
}
