% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{trig_sup}
\alias{trig_sup}
\title{Upper triangular Matrix/Array}
\usage{
trig_sup(A, transp = FALSE, diag = TRUE)
}
\arguments{
\item{A}{a array or a squared matrix}

\item{transp}{boolean, indicate if we need a transposition or not.}

\item{diag}{boolean,  if True, diagonal is not used.}
}
\value{
a array or a squared matrix, with only upper-triangular coefficients with non-zero values
}
\description{
Upper triangular Matrix/Array
}
