% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UpdateParameters.R
\name{update_beta_bayesian}
\alias{update_beta_bayesian}
\title{Update of bayesian parameter beta}
\usage{
update_beta_bayesian(params)
}
\arguments{
\item{params}{list of parameters of the model}
}
\value{
params with beta updated
}
\description{
Update of bayesian parameter beta
}
