% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UpdateParameters.R
\name{update_eta_bayesian}
\alias{update_eta_bayesian}
\title{Update of bayesian parameter eta}
\usage{
update_eta_bayesian(A, params)
}
\arguments{
\item{A}{an array of dim=c(N,N,V)}

\item{params}{list of parameters of the model}
}
\value{
params with eta updated
}
\description{
Update of bayesian parameter eta
}
