% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UpdateParameters.R
\name{update_tau_bayesian}
\alias{update_tau_bayesian}
\title{Update of bayesian parameter tau}
\usage{
update_tau_bayesian(A, params, eps_conv = 1e-04)
}
\arguments{
\item{A}{an array of dim=c(N,N,V)}

\item{params}{list of parameters of the model}

\item{eps_conv}{parameter of convergence.}
}
\value{
params with tau updated
}
\description{
Update of bayesian parameter tau
}
