% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UpdateParameters.R
\name{update_theta_bayesian}
\alias{update_theta_bayesian}
\title{Update of bayesian parameter theta}
\usage{
update_theta_bayesian(params)
}
\arguments{
\item{params}{list of parameters of the model}
}
\value{
params with theta updated
}
\description{
Update of bayesian parameter theta
}
