% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UpdateParameters.R
\name{update_u_bayesian}
\alias{update_u_bayesian}
\title{Update of bayesian parameter u}
\usage{
update_u_bayesian(A, params)
}
\arguments{
\item{A}{an array of dim=c(N,N,V)}

\item{params}{list of parameters of the model}
}
\value{
params with u updated
}
\description{
Update of bayesian parameter u
}
