% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UpdateParameters.R
\name{update_xi_bayesian}
\alias{update_xi_bayesian}
\title{Update of bayesian parameter xi}
\usage{
update_xi_bayesian(A, params)
}
\arguments{
\item{A}{an array of dim=c(N,N,V)}

\item{params}{list of parameters of the model}
}
\value{
params with xi updated
}
\description{
Update of bayesian parameter xi
}
