% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VBEM.R
\name{VBEM_step}
\alias{VBEM_step}
\title{Variational Bayes Expectation Maximization}
\usage{
VBEM_step(A, params, alternate = TRUE, eps_conv = 0.001)
}
\arguments{
\item{A}{an array of dim=c(N,N,V)}

\item{params}{list of parameters of the model}

\item{alternate}{boolean indicated if we put an M-step after each part of the E-step, after u optimization and after tau optimization. If not, we optimize u and tau and after the M-step is made.}

\item{eps_conv}{parameter of convergence for tau.}
}
\value{
params with updated parameters.
}
\description{
Variational Bayes Expectation Maximization
}
